move transforms to product
